chore: remove unused rocket code
