+++ nym-api/src/ecash/api_routes/handlers.rs
@@ -1,10 +1,10 @@
 // Copyright 2023-2024 - Nym Technologies SA <contact@nymtech.net>
 // SPDX-License-Identifier: GPL-3.0-only
 
-use crate::ecash::api_routes::aggregation_axum::aggregation_routes;
-use crate::ecash::api_routes::issued_axum::issued_routes;
-use crate::ecash::api_routes::partial_signing_axum::partial_signing_routes;
-use crate::ecash::api_routes::spending_axum::spending_routes;
+use crate::ecash::api_routes::aggregation::aggregation_routes;
+use crate::ecash::api_routes::issued::issued_routes;
+use crate::ecash::api_routes::partial_signing::partial_signing_routes;
+use crate::ecash::api_routes::spending::spending_routes;
 use crate::ecash::state::EcashState;
 use crate::support::http::state::AppState;
 use axum::Router;
