+++ nym-api/src/ecash/api_routes/mod.rs
@@ -7,8 +7,8 @@ mod helpers;
 // pub(crate) mod partial_signing;
 // pub(crate) mod spending;
 
-pub(crate) mod aggregation_axum;
+pub(crate) mod aggregation;
 pub(crate) mod handlers;
-pub(crate) mod issued_axum;
-pub(crate) mod partial_signing_axum;
-pub(crate) mod spending_axum;
+pub(crate) mod issued;
+pub(crate) mod partial_signing;
+pub(crate) mod spending;
