+++ nym-api/src/node_status_api/mod.rs
@@ -16,59 +16,12 @@ pub(crate) mod handlers;
 pub(crate) mod helpers;
 pub(crate) mod models;
 pub(crate) mod reward_estimate;
-// pub(crate) mod routes_deprecated;
 pub(crate) mod uptime_updater;
 pub(crate) mod utils;
 
 pub(crate) const FIFTEEN_MINUTES: Duration = Duration::from_secs(900);
 pub(crate) const ONE_HOUR: Duration = Duration::from_secs(3600);
 pub(crate) const ONE_DAY: Duration = Duration::from_secs(86400);
-
-// pub(crate) fn node_status_routes(
-//     settings: &OpenApiSettings,
-//     enabled: bool,
-// ) -> (Vec<Route>, OpenApi) {
-//     if enabled {
-//         openapi_get_routes_spec![
-//             settings: routes_deprecated::gateway_report,
-//             routes_deprecated::gateway_uptime_history,
-//             routes_deprecated::gateway_core_status_count,
-//             routes_deprecated::mixnode_report,
-//             routes_deprecated::mixnode_uptime_history,
-//             routes_deprecated::mixnode_core_status_count,
-//             routes_deprecated::get_mixnode_status,
-//             routes_deprecated::get_mixnode_reward_estimation,
-//             routes_deprecated::compute_mixnode_reward_estimation,
-//             routes_deprecated::get_mixnode_stake_saturation,
-//             routes_deprecated::get_mixnode_inclusion_probability,
-//             routes_deprecated::get_mixnode_avg_uptime,
-//             routes_deprecated::get_gateway_avg_uptime,
-//             routes_deprecated::get_mixnode_inclusion_probabilities,
-//             routes_deprecated::get_mixnodes_detailed,
-//             routes_deprecated::get_mixnodes_detailed_unfiltered,
-//             routes_deprecated::get_rewarded_set_detailed,
-//             routes_deprecated::get_active_set_detailed,
-//             routes_deprecated::get_gateways_detailed,
-//             routes_deprecated::get_gateways_detailed_unfiltered,
-//             routes_deprecated::unstable::mixnode_test_results,
-//             routes_deprecated::unstable::gateway_test_results,
-//             routes_deprecated::submit_gateway_monitoring_results,
-//             routes_deprecated::submit_node_monitoring_results,
-//         ]
-//     } else {
-//         // in the minimal variant we would not have access to endpoints relying on existence
-//         // of the network monitor and the associated storage
-//         openapi_get_routes_spec![
-//             settings: routes_deprecated::get_mixnode_status,
-//             routes_deprecated::get_mixnode_stake_saturation,
-//             routes_deprecated::get_mixnode_inclusion_probability,
-//             routes_deprecated::get_mixnode_inclusion_probabilities,
-//             routes_deprecated::get_mixnodes_detailed,
-//             routes_deprecated::get_rewarded_set_detailed,
-//             routes_deprecated::get_active_set_detailed,
-//         ]
-//     }
-// }
 
 /// Spawn the node status cache refresher.
 ///
