--- nym-api/src/nym_contract_cache/legacy_helpers.rs
@@ -1,2 +1,0 @@
-// Copyright 2024 - Nym Technologies SA <contact@nymtech.net>
-// SPDX-License-Identifier: GPL-3.0-only
